requirements.txt
appnope==0.1.0
certifi==2017.4.17
chardet==3.0.4
decorator==4.0.11
Django==1.11.2
django-extensions==1.7.9
djangorestframework==3.6.3
google-api-python-client==1.6.2
httplib2==0.10.3
idna==2.5
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
oauth2client==4.1.1
olefile==0.44
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
pyasn1==0.2.3
pyasn1-modules==0.0.9
Pygments==2.2.0
pytz==2017.2
requests==2.18.1
rsa==3.4.2
selenium==3.4.3
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.21.1
wcwidth==0.1.7
